+++ packages/mongo/src/index.ts
@@ -1,9 +1,3 @@
 // New memory system - preferred API
 export * from "./memory";
 export * from "./providers";
-
-// Legacy exports - deprecated
-/**
- * @deprecated Use the new memory system providers instead
- */
-export * from "./mongo";